Desktop/projects/Fitapp-fixierun-full/FitApp-Consolidated/apps/backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
gunicorn==21.2.0

# Database
pymongo==4.6.1
motor==3.3.2
Flask-MongoEngine==1.0.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Web3 Integration
web3==6.15.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.1
geopy==2.4.1
requests==2.31.0
tinydb==4.8.0

# Testing
pytest==7.4.4
pytest-dotenv==0.5.2
unittest2==1.1.0

# Development Tools
black==24.1.1
mypy==1.8.0
flake8==7.0.0
pre-commit==3.6.0